refactor(store-core): extract auth/connection handlers into handlers/auth.ts (audit P2-3)
Pure byte-identical move behind the ./index barrel. Extracts the auth +
connection family from handlers/index.ts into a new handlers/auth.ts:

- auth_ok / auth_fail / key_exchange_ok / server_mode (with the ServerMode +
  AuthOkWebFeatures + AuthOkPayload types, DEFAULT_WEB_FEATURES and
  parsePositiveFiniteNumber helpers)
- the connect-time auth_bootstrap + tunnel_url_changed pair (with the shared
  asWssUrl helper)
- token_rotated + pair_fail (with PAIR_FAIL_MESSAGES + PairFailPayload)

handleAuthOk's import-back of handleAvailablePermissionModes + PermissionMode
moves with it to ./auth (no cycle — ./permission never imports ./auth/./index);
index.ts drops that import plus the now-unused parseEnumField /
parseRawStringField from its ./_shared import. Re-exported via `export * from
'./auth'` so the barrel's public surface is unchanged.

index.ts 1109 -> 706 lines. Move proven byte-identical by reversal; tsc clean;
1574 store-core tests pass (790 handler tests).
